+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## unreleased
 
 * Updated `nalgebra` to 0.10.
+* Rename traits `HasSVD` to `SVD` and `HasEigensystem` to `Eigensystem`.
 * Added `Solve` trait for solving a linear matrix equation.
 * Added `Inverse` for computing the multiplicative inverse of a matrix.
 * Added `Cholesky` for decomposing a positive-definite matrix.